+++ fx/use-reduced-motion.d.ts
@@ -1,0 +1,2 @@
+/** @returns {boolean} */
+export declare function useReducedMotion(): boolean;